+++ lib/factory/browser_factory.js
@@ -6,8 +6,11 @@ const AssertionFactory = require('./assertion_factory');
 const FactoryCore = require('./factory_core');
 
 const mixins = [
+    require("../mixins/browser_setup"),
+    require("../mixins/browser_queries"),
     require("../mixins/browser_base_mixin"),
-    require("../mixins/browser_mixin")
+    require("../mixins/browser_mixin"),
+    require("../mixins/browser_navigation")
 ];
 
 const components = {
