+++ lib/mixins/browser_base_mixin.js
@@ -3,50 +3,12 @@
 
 const ErrorFactory = require('../errors/error_factory');
 
-function pageLog(log) {
-    console[log._type](log._text); // eslint-disable-line
-}
-
 module.exports = function BrowserBaseMixin(s) {
     return class BrowserBase extends s {
-        constructor(page, settings) {
-            super(page, settings);
-            this._loaded = false;
-            if(settings.log) {
-                this.page.on("console", pageLog);
-            }
-        }
 
         evaluate(cb, ...args) {
             this._failIfNotLoaded();
             return this.page.evaluate(cb, ...args);
-        }
-
-        query(selector, optionalSelector) {
-            this._failIfNotLoaded();
-            if(optionalSelector) {
-                return this._subQuery(selector, optionalSelector);
-            }
-            if(typeof selector === 'string') {
-                return this.page.$(selector);
-            } else return Promise.resolve(selector);
-        }
-
-        queryAll(selector, optionalSelector) {
-            this._failIfNotLoaded();
-            if(optionalSelector) {
-                return this._subQueryAll(selector, optionalSelector);
-            }
-            if(typeof selector === 'string') {
-                return this.page.$$(selector);
-            } else if(!Array.isArray(selector)) {
-                return Promise.resolve([selector]);
-            } else return Promise.resolve(selector);
-        }
-
-        queryXPath(xPath) {
-            this._failIfNotLoaded();
-            return this.page.$x(xPath);
         }
 
         class(selector) {
@@ -93,29 +55,6 @@ module.exports = function BrowserBaseMixin(s) {
                 const error = ErrorFactory.generateQueryError(`Element "${selector}" not found when trying to get styles.`);
                 return Promise.reject(error);
             });
-        }
-
-        _subQueryXpath(parent, selector) {
-            if(selector[1] === "/") selector = selector.slice(1);
-            if(selector[0] === "/")selector = `.${ selector}`;
-            if(!parent || !parent.$) return Promise.reject(new Error("Invalid parent element for query"));
-            return parent.$x(selector);
-        }
-
-        _subQuery(parent, selector) {
-            if(!parent || !parent.$) return Promise.reject(new Error("Invalid parent element for query"));
-            return parent.$(selector);
-        }
-
-        _subQueryAll(parent, selector) {
-            if(!parent || !parent.$) return Promise.reject(new Error("Invalid parent element for queryAll"));
-            return parent.$$(selector);
-        }
-
-        _failIfNotLoaded() {
-            if(!this._loaded) {
-                throw ErrorFactory.generateFatalError(`Cannot perform action before opening a page.`);
-            }
         }
     };
 };
