+++ lib/mixins/browser_mixin.js
@@ -1,36 +1,14 @@
 /* global WendigoUtils */
 "use strict";
 
-const path = require('path');
 const ErrorFactory = require('../errors/error_factory');
 const utils = require('../utils');
-const config = require('../../config');
-
-const injectionScriptsPath = config.injectionScripts.path;
-const injectionScripts = config.injectionScripts.files;
 
 module.exports = function BrowserMixin(s) {
     return class BrowserMixin extends s {
 
         constructor(page, settings) {
             super(page, settings);
-        }
-
-        open(url) {
-            return this.page.goto(url).then(() => {
-                return this._afterPageLoad();
-            }).catch(() => {
-                return Promise.reject(ErrorFactory.generateFatalError(`Failed to open ${url}.`));
-            });
-        }
-
-        close() {
-            if(this._loaded === false) return Promise.resolve();
-            this._loaded = false;
-            this._originalHtml = undefined;
-            return this.page.close().catch(() => {
-                return Promise.reject(ErrorFactory.generateFatalError(`Failed to close browser.`));
-            });
         }
 
         text(selector) {
@@ -250,43 +228,6 @@ module.exports = function BrowserMixin(s) {
                 const error = ErrorFactory.generateQueryError(`Element "${selector}" not found when trying to get selected options.`);
                 return Promise.reject(error);
             });
-        }
-
-        back() {
-            this._failIfNotLoaded();
-            return this.page.goBack().then(() => {
-                return this._afterPageLoad();
-            });
-        }
-
-        forward() {
-            this._failIfNotLoaded();
-            return this.page.goForward().then(() => {
-                return this._afterPageLoad();
-            });
-        }
-
-        refresh() {
-            this._failIfNotLoaded();
-            return this.page.reload().then(() => {
-                return this._afterPageLoad();
-            });
-        }
-
-        _afterPageLoad() {
-            return this.page.content().then((content) => {
-                this._originalHtml = content;
-                return this._addJsScripts().then(() => {
-                    this._loaded = true;
-                });
-            });
-        }
-
-        _addJsScripts() {
-            const promises = injectionScripts.map((s) => {
-                return this.page.addScriptTag({path: path.join(injectionScriptsPath, s)});
-            });
-            return Promise.all(promises);
         }
     };
 };
