+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.5.0",
   "description": "A proper monster for front-end automated testing",
   "engines": {
-    "node": ">=8.9.4"
+    "node": ">=8.0.0"
   },
   "main": "lib/wendigo.js",
   "scripts": {
